Resolves #284
